backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic-settings==2.2.1
python-jose==3.3.0
boto3==1.34.34
redis==5.0.1
python-dotenv==1.0.1
python-multipart==0.0.9
pillow==10.2.0
torch==2.2.2
torchvision==0.17.2
ultralytics
email-validator==2.1.1
requests==2.32.3
